Dockerfile: Use C++17 to build etcd
Also, fix #1054 where etcd event type needs to be cast to int to be
displayed.

+++ src/core/nixl_listener.cpp
@@ -409,7 +409,7 @@ public:
                 invalidated_agents.push_back(agent_name);
             } else {
                 NIXL_ERROR << "Watcher for " << event.kv().key() << " received unexpected event from etcd: "
-                           << event.event_type();
+                           << static_cast<int>(event.event_type());
             }
         };
